refactor(canvas): extract logic into custom hooks for better separation of concerns
- refactor(ui): move Drag & Drop logic to 'useDiagramDnD'.
- refactor(ui): centralization context menu logic in 'useDiagramMenus'.
- refactor(ui): extract keyboard shortcuts to 'useKeyboardShortcuts'.
- refactor(ui): move edge styling calculation to 'useEdgeStyling'.
- cleanup: simplify DiagramCanvas component significantly.
