+++ frontend/src/features/diagram/hooks/useKeyboardShortcuts.ts
@@ -1,0 +1,33 @@
+import { useEffect } from "react";
+import { useDiagramStore } from "../../../store/diagramStore";
+
+export const useKeyboardShortcuts = () => {
+  useEffect(() => {
+    const handleKeyDown = (e: KeyboardEvent) => {
+      if (
+        e.target instanceof HTMLInputElement ||
+        e.target instanceof HTMLTextAreaElement
+      )
+        return;
+
+      const temporal = useDiagramStore.temporal.getState();
+
+      if ((e.ctrlKey || e.metaKey) && e.key === "z") {
+        e.preventDefault();
+        if (e.shiftKey) {
+          temporal.redo();
+        } else {
+          temporal.undo();
+        }
+      }
+
+      if ((e.ctrlKey || e.metaKey) && e.key === "y") {
+        e.preventDefault();
+        temporal.redo();
+      }
+    };
+
+    window.addEventListener("keydown", handleKeyDown);
+    return () => window.removeEventListener("keydown", handleKeyDown);
+  }, []);
+};
